refactor: organize creator profiles by file
